tweak

+++ src/NetEventLoop.cpp
@@ -123,7 +123,7 @@ void NetEventLoop::OnPollIn(int fd)
 	//If we got listener
 	if (listener)
 		//for each one
-		for (int i = 0; i < len && i < MaxMultipleReceivingMessages; i++)
+		for (size_t i = 0; int(i) < len && i < MaxMultipleReceivingMessages; i++)
 			//double check
 			if (messages[i].msg_len)
 				//Run callback
